First light
This project is now roughly one year old, and more or less ready for public display. The tool slash wizard runs decently when launched as a script (on some bad hair days still a bit glitchy though). Next steps will be building/freezing/deploying it as a proper app, and improving stability.